+++ src/lte/model/lte-rrc-header.h
@@ -72,14 +72,14 @@ protected:
    * \param bandwidth Bandwidth in RBs: 6, 15, 25, 50, 75, 100
    * \returns ENUMERATED value: 0, 1, 2, 3, 4, 5
    */
-  int BandwidthToEnum (uint8_t bandwidth) const;
+  int BandwidthToEnum (uint16_t bandwidth) const;
   /**
    * Convert from ENUMERATED value to bandwidth (in RBs)
    *
    * \param n ENUMERATED value: 0, 1, 2, 3, 4, 5
    * \returns bandwidth Bandwidth in RBs: 6, 15, 25, 50, 75, 100
    */
-  uint8_t EnumToBandwidth (int n) const;
+  uint16_t EnumToBandwidth (int n) const;
 
 
   // Serialization functions
